Renamed package structure to com.camacho.simpleannotation
