+++ src/main/java/com/camacho/simpleannotation/exceptions/ClassGenerationException.java
@@ -1,0 +1,11 @@
+package com.camacho.simpleannotation.exceptions;
+
+public class ClassGenerationException extends RuntimeException{
+    public ClassGenerationException () {
+        super();
+    }
+    public ClassGenerationException (Throwable t) {
+        super(t);
+    }
+
+}
