--- src/main/java/com/example/simpleannotation/exceptions/ClassGenerationException.java
@@ -1,11 +1,0 @@
-package com.example.simpleannotation.exceptions;
-
-public class ClassGenerationException extends RuntimeException{
-    public ClassGenerationException () {
-        super();
-    }
-    public ClassGenerationException (Throwable t) {
-        super(t);
-    }
-
-}
